pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Difference calculator is the educational project on the hexlet"
authors = ["Aleksandr Chepaikin <alexgreendev@gmail.com>"]
packages = [
  { include = "gendiff" },
]
classifiers = [
    "Topic :: Education"
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest = "^6.2.2"
pytest-asyncio = "^0.14.0"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
